% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{context}
\alias{context}
\title{See Terms in Context}
\usage{
context(x, regex, file = "out.html", n.before = 1, n.after = n.before,
  color = "lightgreen", open = TRUE, text.var = TRUE,
  ignore.case = FALSE, ...)
}
\arguments{
\item{x}{A data.frame with a text column.}

\item{regex}{A regular expression that will be highlighted.}

\item{file}{A file to output the .html output to.}

\item{n.before}{The number of rows before an occurrence to show.}

\item{n.after}{The number of rows after an occurrence to show.}

\item{color}{The highlight color.}

\item{open}{logical.  If \code{TRUE} the created .html file will attempt to be opened.}

\item{text.var}{The name of the text variable. If TRUE split_word tries to detect the text column with words.}

\item{ignore.case}{logical.  If \code{TRUE} case sensitivity is ignored.}

\item{\ldots}{ignored.}
}
\value{
Generates an .html file with highlighted terms in context.
}
\description{
Examine regex term selections in the context of n rows before and after.
}

