% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textcode-package.R
\docType{package}
\name{textcode}
\alias{package-textcode}
\alias{textcode}
\alias{textcode-package}
\title{Tools for Qualitative Text Coding}
\description{
Tools for assisting in qualitative text coding and analysis.  Tools to:
export text for qualitative coding, import coding schemes, and reshape the
imports into an analysis ready format.
}

